Create a file for pylint.reporters.BaseWritter
